+++ packages/uploader/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.1.50 (October 30, 2024)
+
+- chore: remove useless deps
+
+## 0.1.49 (October 30, 2024)
+
+- chore: re-bump version
+
+## 0.1.48 (October 30, 2024)
+
+- chore: split @blocklet/uploader lib
+
+## 0.1.47 (October 29, 2024)
+
+- feat: support aspect ratio before upload
+- feat: support image editor props
+
 ## 0.1.46 (October 22, 2024)
 
 - chore: update deps